refactor: split creation and validation of NumberBallSet
NumberBallSet의 검증 로직을 validator 클래스로 분리
